Add annotation groups to the AnnotationSelector panel
This groups the annotation list in the AnnotationSelector panel by
"group" name.  In this context, the group the annotation belongs to is a
list of one or more styles applied to at least one element.  In
addition, if an element in an annotation has no style name, then it is
listed as in the "null" group which gets displayed into a single group
named "Other".

This depends on https://github.com/girder/large_image/pull/276